Not allow out-of-bounds moves
Added a message that triggers on an out-of-bounds move. If out-of-bounds
move is inputed askForPlay will run again and update currentMove
accordingly.

Pretty messy with all the comments. I'll clean it up when I figure out
how to disallow illegal moves (moves on spaces that are already claimed
by X or O).