Not allow illegal moves
Whew, this one took a while! Illegal moves are checked for and if one's
trying to be done, alert the player and reprompt them for a move. The
clincher was "board[v] !=0 && v == currentMove". Basically, check to see
if a space is already occupied and the player is trying to move there.
If so, let them know and run askForPlay again. Good placement of the
trunNumber increment by sts is key for this not to break a lot of other
stuff.

See line 37.