Fix bugs for repeated moves that are not allowed
Had to rewrite how currentMove is checked. Moved currentMove's
declaration (hope that doesn't bug anything) and created another
variable, illegalMove, that is a bool to trigger the exit from the for
loop. Wait, I might have allowed for repeated out-of-bounds moves.
Speaking of, I updated the out-of-bounds message. Sometimes you eat the
bar, and sometimes the bar eats you.