Fixed?
Still not sure about the for loop starting on 22 and how it's handling
currentMove being > 9. It still seems to work and so did the last
commit, which is not what I expected. I still can't figure out how to
debug with inputs on vscode, so I can't really get that much info.
Regardless, it works.