Check for occupied spaces
Added a loop to check if the currentMove (moveInt in the function) would
go in an occupied space which flips a bool, legalMove. Then, a switch
will check legalMove. If it's false then we ask for another input and
run executePlayerMove again which will do the check and reprompt if it
is again an occupied space.