Alternate way to check for cat's game
Instead of relying on the turnNumber count to generate cat's game we
check to see if all spaces are occupied. I added the loop to check in
checkForWin() and if will return 3. Up in main I made a switch check for
result. If checkForWin() gets all the way through without spotting a win
or cat's game (case 0) we keep playing and increment the turnNumber. 1
and 2 are winners. 3 is cat's game.